+++ packages/navbar/src/LeftPanelSection/AuthedNav/ToolBox/index.tsx
@@ -3,13 +3,16 @@ import { PROPERTY_REPORT_LINK, REPAYMENT_CALCULATOR_LINK } from './../../../cons
 import { SidebarNav } from './../../../SidebarNav';
 import { ListItem } from './index.style';
 
-export const ToolBoxSection = () => (
+export interface ToolBoxSectionProps {
+  params?: string;
+}
+export const ToolBoxSection = ({ params = '' }: ToolBoxSectionProps) => (
   <ListItem>
     <SidebarNav.Section title="Your toolbox">
-      <SidebarNav.Item href={PROPERTY_REPORT_LINK} labelText="Property report">
+      <SidebarNav.Item href={`${PROPERTY_REPORT_LINK}${params}`} labelText="Property report">
         Property report
       </SidebarNav.Item>
-      <SidebarNav.Item href={REPAYMENT_CALCULATOR_LINK} labelText="Repayments calculator">
+      <SidebarNav.Item href={`${REPAYMENT_CALCULATOR_LINK}${params}`} labelText="Repayments calculator">
         Repayments calculator
       </SidebarNav.Item>
     </SidebarNav.Section>