+++ packages/navbar/src/LeftPanelSection/AuthedNav/index.tsx
@@ -13,6 +13,7 @@ interface AuthedNavDetails {
   applicationSfid?: string;
   transactionType?: string;
   applicationNumber?: number;
+  params?: string;
 }
 
 export const AuthedNav = ({
@@ -21,6 +22,7 @@ export const AuthedNav = ({
   applicationSfid,
   transactionType,
   applicationNumber,
+  params = '',
 }: AuthedNavDetails) => {
   const primaryApplicant = applicants.find((applicant) => applicant.type === 'primary');
   const primaryApplicantSfid = (primaryApplicant && primaryApplicant.sfid) || '';
@@ -39,7 +41,7 @@ export const AuthedNav = ({
     case APPLICATION_STAGES[0]: {
       return (
         <SidebarNav.Section>
-          <SidebarNav.Item href={profileLink} labelText="Your profile">
+          <SidebarNav.Item href={`${profileLink}${params}`} labelText="Your profile">
             Your profile
           </SidebarNav.Item>
           <FunnelTwoSection
@@ -47,15 +49,16 @@ export const AuthedNav = ({
             transactionType={transactionType}
             numberOfApplicants={numberOfApplicants}
             applicationNumber={applicationNumber}
+            params={params}
           />
-          <ToolBoxSection />
+          <ToolBoxSection params={params} />
         </SidebarNav.Section>
       );
     }
     case APPLICATION_STAGES[1]: {
       return (
         <SidebarNav.Section>
-          <SidebarNav.Item href={profileLink} labelText="Your profile">
+          <SidebarNav.Item href={`${profileLink}${params}`} labelText="Your profile">
             Your profile
           </SidebarNav.Item>
           <FunnelTwoSection
@@ -63,19 +66,21 @@ export const AuthedNav = ({
             transactionType={transactionType}
             numberOfApplicants={numberOfApplicants}
             applicationNumber={applicationNumber}
+            params={params}
           />
           <FunnelThreeSection
             applicants={applicants}
             applicationSfid={applicationSfid}
             applicationNumber={applicationNumber}
+            params={params}
           />
-          <SidebarNav.Item href={verifyIdLink} labelText="Verify your ID">
+          <SidebarNav.Item href={`${verifyIdLink}${params}`} labelText="Verify your ID">
             Verify your ID
           </SidebarNav.Item>
-          <SidebarNav.Item href={documentLink} labelText="Upload your documents">
+          <SidebarNav.Item href={`${documentLink}${params}`} labelText="Upload your documents">
             Upload your documents
           </SidebarNav.Item>
-          <SidebarNav.Item href={dashOverviewLink} labelText="Your dashboard">
+          <SidebarNav.Item href={`${dashOverviewLink}${params}`} labelText="Your dashboard">
             Your dashboard
           </SidebarNav.Item>
           <ToolBoxSection />
@@ -85,7 +90,7 @@ export const AuthedNav = ({
     case APPLICATION_STAGES[2]: {
       return (
         <SidebarNav.Section>
-          <SidebarNav.Item href={profileLink} labelText="Your profile">
+          <SidebarNav.Item href={`${profileLink}${params}`} labelText="Your profile">
             Your profile
           </SidebarNav.Item>
           <FunnelTwoSection
@@ -99,35 +104,35 @@ export const AuthedNav = ({
             applicationSfid={applicationSfid}
             applicationNumber={applicationNumber}
           />
-          <SidebarNav.Item href={verifyIdLink} labelText="Verify your ID">
+          <SidebarNav.Item href={`${verifyIdLink}${params}`} labelText="Verify your ID">
             Verify your ID
           </SidebarNav.Item>
-          <SidebarNav.Item href={summaryLink} labelText="View your summary">
+          <SidebarNav.Item href={`${summaryLink}${params}`} labelText="View your summary">
             View your summary
           </SidebarNav.Item>
-          <SidebarNav.Item href={lenderSelectionLink} labelText="Choose your loan">
+          <SidebarNav.Item href={`${lenderSelectionLink}${params}`} labelText="Choose your loan">
             Choose your loan
           </SidebarNav.Item>
-          <SidebarNav.Item href={documentLink} labelText="Upload your documents">
+          <SidebarNav.Item href={`${documentLink}${params}`} labelText="Upload your documents">
             Upload your documents
           </SidebarNav.Item>
-          <SidebarNav.Item href={dashOverviewLink} labelText="Your dashboard">
+          <SidebarNav.Item href={`${dashOverviewLink}${params}`} labelText="Your dashboard">
             Your dashboard
           </SidebarNav.Item>
-          <ToolBoxSection />
+          <ToolBoxSection params={params} />
         </SidebarNav.Section>
       );
     }
     default: {
       return (
         <SidebarNav.Section>
-          <SidebarNav.Item href={SEARCH_LOAN_LINK} labelText="Start a new loan">
+          <SidebarNav.Item href={`${SEARCH_LOAN_LINK}${params}`} labelText="Start a new loan">
             Start a new loan
           </SidebarNav.Item>
-          <SidebarNav.Item href={MANAGE_APPLICATION_LINK} labelText="My dashboard">
+          <SidebarNav.Item href={`${MANAGE_APPLICATION_LINK}${params}`} labelText="My dashboard">
             My dashboard
           </SidebarNav.Item>
-          <SidebarNav.Item href={PROPERTY_REPORT_LINK} labelText="My property reports">
+          <SidebarNav.Item href={`${PROPERTY_REPORT_LINK}${params}`} labelText="My property reports">
             My property reports
           </SidebarNav.Item>
         </SidebarNav.Section>