+++ packages/navbar/src/LeftPanelSection/index.tsx
@@ -7,9 +7,10 @@ export interface LeftPanelSectionProps {
   isAuthenticated?: boolean;
   application?: Application;
   applicants?: Applicant[];
+  params?: string;
 }
 
-export const LeftPanelSection = ({ isAuthenticated, application, applicants }: LeftPanelSectionProps) => {
+export const LeftPanelSection = ({ isAuthenticated, application, applicants, params = '' }: LeftPanelSectionProps) => {
   if (isAuthenticated && application) {
     const { stage, sfid, type, number } = application;
     return (
@@ -19,13 +20,14 @@ export const LeftPanelSection = ({ isAuthenticated, application, applicants }: L
         transactionType={type}
         applicants={applicants}
         applicationNumber={number}
+        params={params}
       />
     );
   }
 
   // user who has an account but doesn't have any application/opportunity yet
   if (isAuthenticated && !application) {
-    return <AuthedNav />;
+    return <AuthedNav params={params} />;
   }
-  return <UnauthedNav />;
+  return <UnauthedNav params={params} />;
 };